adding tests via modules
