+++ projects/test-servers/shared/modules/initRequest/server.ts
@@ -1,0 +1,18 @@
+import { Module } from '../../../../core/server/module.js'
+
+/**
+ * in next-server, we need to copy the same module...
+ *
+ * Here is the [file to edit](../../../next-server/src/pages/api/[...remult].ts)
+ *
+ * I would love to have only THIS module to test, all server are working except next-server...!
+ * Maybe the shared modules should be in next-server!
+ */
+export const initRequestModule = new Module({
+  key: 'init-request-module',
+  async initRequest(_, { req }) {
+    if (req.headers.get('remult-test-crash') === 'yes') {
+      throw new Error('test crash')
+    }
+  },
+})
