+++ projects/tests/backend-tests/all-server-tests.ts
@@ -575,6 +575,29 @@ export function allServerTests(
       }
     }),
   )
+
+  describe('Module', () => {
+    describe('initRequest', () => {
+      it(
+        'should read generic request headers and crash',
+        withRemultForTest(async () => {
+          try {
+            await axios.get(remult.apiClient.url + '/me', {
+              headers: { 'remult-test-crash': 'yes' },
+            })
+            expect('to never').toBe('be here')
+          } catch (error) {
+            if (error instanceof Error) {
+              expect(error.message).toMatchInlineSnapshot(
+                `"Request failed with status code 400"`,
+              )
+            }
+          }
+        }),
+      )
+    })
+  })
+
   async function create3Tasks() {
     const taskRepo = remult.repo(Task)
