+++ projects/tests/backend-tests/test-fastify-server.spec.ts
@@ -6,6 +6,7 @@ import { Task } from '../../test-servers/shared/Task.js'
 import { Remult, remult } from '../../core'
 import { RemultAsyncLocalStorage } from '../../core/src/context.js'
 import { allServerTests } from './all-server-tests.js'
+import { initRequestModule } from '../../test-servers/shared/modules/initRequest/server.js'
 
 describe('test fastify server', async () => {
   let destroy: () => Promise<void>
@@ -17,6 +18,7 @@ describe('test fastify server', async () => {
       const api = remultApi({
         entities: [Task],
         admin: true,
+        modules: [initRequestModule],
       })
       await app.register(api)
       app.get('/api/test', async (req, res) => {
