+++ projects/tests/backend-tests/test-koa-server.spec.ts
@@ -5,6 +5,7 @@ import { afterAll, beforeAll, describe, expect, it } from 'vitest'
 import { Task } from '../../test-servers/shared/Task.js'
 import { RemultAsyncLocalStorage } from '../../core/src/context.js'
 import { allServerTests } from './all-server-tests.js'
+import { initRequestModule } from '../../test-servers/shared/modules/initRequest/server.js'
 
 describe.skipIf(process.env['SKIP_KOA'])('test koa server', async () => {
   let destroy: () => Promise<void>
@@ -16,6 +17,7 @@ describe.skipIf(process.env['SKIP_KOA'])('test koa server', async () => {
       const api = createRemultServer({
         entities: [Task],
         admin: true,
+        modules: [initRequestModule],
       })
       app.use(bodyParser())
       app.use(async (ctx, next) => {
